==> manuscript/Rmd/02-04-simple-logit.Rmd <==
---
title: "02-04-Simple-Logit-Models"
author: "gmlang"
date: "April 3, 2015"
output: pdf_document
---

```{r setup}
library(knitr)
opts_chunk$set(comment = "", warning = FALSE, message = FALSE, tidy = FALSE,
               echo = TRUE, fig.width = 6, fig.height = 6, dev = 'png')
options(width = 100, scipen = 5, digits = 5)
```

## Simple Logit Models

First, let's load the data we saved from section 3.2.

```{r}
proj_path = "~/score-loan-applicants"
data_path = file.path(proj_path, 'data')
file_path = file.path(data_path, 'cleaned-02.rda')
load(file_path)
```

Next, we fit a simple logit model regressing the target against each predictor separately and extract the parameter estimates and their p-values. We do that using a for loop.

```{r}
predictors = c(iv_cat_strong, iv_cat_weak, iv_cat_none, 
               iv_con_strong, iv_con_weak, iv_con_none)
simple_est = data.frame()
for (var in predictors) {
        f = formula(paste0("bad ~ ", var))
        fit = glm(f, data=upl, family=binomial)
        simple_est = rbind(simple_est, rbind(summary(fit)$coef[2, c("Estimate", "Pr(>|z|)")]))
} 
simple_est = cbind(var=predictors, simple_est)
simple_est$Significant = ifelse(simple_est[["Pr(>|z|)"]] < 0.05, "*", " ")
print(simple_est)
```

We then separate the predictors into two groups based on if their p-values are greater than 0.05 or not. And we save them for ensuing analysis.

```{r}
iv_sig = as.character(simple_est$var[simple_est[["Pr(>|z|)"]] <= 0.05])
print(iv_sig)
iv_non_sig = as.character(simple_est$var[simple_est[["Pr(>|z|)"]] > 0.05])
print(iv_non_sig)
save(iv_sig, iv_non_sig, file=file.path(data_path, 'cleaned-04.rda'))
```
